+++ AutomatedLab/AutomatedLab.psm1
@@ -1581,6 +1581,9 @@ function Remove-Lab
 
                 Write-ScreenInfo -Message 'Removing entries in the hosts file'
                 Clear-HostFile -Section $Script:data.Name -ErrorAction SilentlyContinue
+
+                Write-ScreenInfo -Message 'Removing SSH known hosts'
+                UnInstall-LabSshKnownHost
             }
 
             Write-ScreenInfo -Message 'Removing virtual networks'
@@ -2828,15 +2831,6 @@ function Install-LabSoftwarePackage
     Write-PSFMessage -Message "Starting background job for '$($parameters.ActivityName)'"
 
     $parameters.ScriptBlock = {
-        if ($PSEdition -eq 'core')
-        {
-            Add-Type -Path '/ALLibraries/core/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-        }
-        elseif ([System.Environment]::OSVersion.Version -ge '6.3')
-        {
-            Add-Type -Path '/ALLibraries/full/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-        }
-
         if ($installParams.Path.StartsWith('\\') -and (Test-Path /ALAzure))
         {
             # Often issues with Zone Mapping
@@ -2858,7 +2852,18 @@ function Install-LabSoftwarePackage
             }
             $installParams.Path = $newPath
         }
-        Install-SoftwarePackage @installParams
+
+        if ($PSEdition -eq 'core' -and $installParams.Contains('AsScheduledJob'))
+        {
+            # Core cannot work with PSScheduledJob module
+            $xmlParameters = ([System.Management.Automation.PSSerializer]::Serialize($installParams, 2)) -replace "`r`n"
+            $b64str = [Convert]::ToBase64String(([Text.Encoding]::Unicode.GetBytes("`$installParams = [System.Management.Automation.PSSerializer]::Deserialize('$xmlParameters'); Install-SoftwarePackage @installParams")))
+            powershell.exe -EncodedCommand $b64str
+        }
+        else
+        {
+            Install-SoftwarePackage @installParams
+        }
     }
 
     $parameters.Add('NoDisplay', $True)
@@ -2868,7 +2873,7 @@ function Install-LabSoftwarePackage
         Write-ScreenInfo -Message "Copying files and initiating setup on '$($ComputerName -join ', ')' and waiting for completion" -NoNewLine
     }
 
-    $job = Invoke-LabCommand @parameters -Variable (Get-Variable -Name installParams) -Function (Get-Command -Name Install-SoftwarePackage)
+    $job = Invoke-LabCommand @parameters -Variable (Get-Variable -Name installParams)
 
     if (-not $AsJob)
     {