+++ AutomatedLab/AutomatedLabADCS.psm1
@@ -1281,26 +1281,14 @@ function Get-LabCertificate
     Write-LogFunctionEntry
 
     $variables = Get-Variable -Name PSBoundParameters
-    $functions = Get-Command -Name Get-Certificate2, Sync-Parameter
 
     foreach ($computer in $ComputerName)
     {
-        Invoke-LabCommand -ActivityName "Adding 'AutomatedLab.Common.dll'" -ComputerName $ComputerName -ScriptBlock {
-            if ($PSEdition -eq 'core')
-            {
-                Add-Type -Path '/ALLibraries/core/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-            }
-            elseif ([System.Environment]::OSVersion.Version -ge '6.3')
-            {
-                Add-Type -Path '/ALLibraries/full/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-            }
-        } -NoDisplay
-
         Invoke-LabCommand -ActivityName 'Exporting certificates' -ComputerName $ComputerName -ScriptBlock {
             Sync-Parameter -Command (Get-Command -Name Get-Certificate2)
             Get-Certificate2 @ALBoundParameters
 
-        } -Variable $variables -Function $functions -PassThru -NoDisplay
+        } -Variable $variables -PassThru -NoDisplay
     }
 
     Write-LogFunctionExit
@@ -1347,18 +1335,6 @@ function Add-LabCertificate
     process
     {
         $variables = Get-Variable -Name PSBoundParameters
-        $functions = Get-Command -Name Add-Certificate2, Sync-Parameter
-
-        Invoke-LabCommand -ActivityName "Adding 'AutomatedLab.Common.dll'" -ComputerName $ComputerName -ScriptBlock {
-            if ($PSEdition -eq 'core')
-            {
-                Add-Type -Path '/ALLibraries/core/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-            }
-            elseif ([System.Environment]::OSVersion.Version -ge '6.3')
-            {
-                Add-Type -Path '/ALLibraries/full/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-            }
-        } -NoDisplay
 
         if ($Path)
         {
@@ -1372,7 +1348,7 @@ function Add-LabCertificate
             Sync-Parameter -Command (Get-Command -Name Add-Certificate2)
             Add-Certificate2 @ALBoundParameters | Out-Null
 
-        } -Variable $variables -Function $functions -PassThru -NoDisplay
+        } -Variable $variables -PassThru -NoDisplay
 
     }